requirements.txt
Django
requests
requests-oauthlib
django-bootstrap-datepicker-plus
django-bootstrap4
